requirements.txt
pyspark==2.4.3
kafka-python==1.4.6
tweepy==4.0.0